Tester/requirements.txt
google-generativeai
python-dotenv
streamlit
pandas
plotly
transformers